Split AI chat runtime tests
